moves stud/lect process and ics function to another file
